消息大小限制改成32位大小

+++ H6Game/H6Game.Base/Base/MessageHandler/OutConnectHandler.cs
@@ -5,7 +5,7 @@ using H6Game.Message;
 
 namespace H6Game.Base
 {
-    [HandlerCMD((int)MessageCMD.GetGateEndPoint)]
+    [HandlerCMD(MessageCMD.GetGateEndPoint)]
     public class OutConnectHandler : AHandler<NetEndPointMessage>
     {
         protected override void Handler(NetEndPointMessage response, int messageId)
